requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.12
openpyxl==3.1.5
pydantic==2.9.2
yfinance>=1.3.0
kiteconnect>=5.1.0
python-dotenv>=1.0.0
